requirements.txt
# Core functionality
PyPDF2>=3.0.1
sentence-transformers>=2.6.1
faiss-cpu>=1.7.4
openai>=1.23.0

# Deep-learning back-end for sentence-transformers & CLIP
torch>=2.2.2
torchvision>=0.17.2
transformers>=4.40.0

# Utilities
pillow>=10.3.0
pytesseract>=0.3.10
nltk>=3.8.1

tqdm>=4.66.4  # progress bars used by sentence-transformers
pytest>=8.2.0  # test runner

numpy